feat: change model launch params from double to decimal
- Update ModelLaunchParams to use decimal? for exact 2-decimal precision
- Add 'm suffix to default parameter literals
- Update DashboardService.TryGetDecimal() with Math.Round(value.GetDecimal(), 2)
- Change all PowerShell scripts to [decimal] param types
- Ensures llama-server compatibility with precise parameter values

+++ Services/ModelLaunchParams.cs
@@ -4,18 +4,18 @@ namespace LlaModem.Services;
 /// Optional parameters passed via HTTP headers to override model launch settings.
 /// </summary>
 public record ModelLaunchParams(
-    double? Temperature,
-    double? TopP,
-    double? TopK,
-    double? MinP,
-    double? PresencePenalty,
-    double? RepetitionPenalty)
+    decimal? Temperature,
+    decimal? TopP,
+    decimal? TopK,
+    decimal? MinP,
+    decimal? PresencePenalty,
+    decimal? RepetitionPenalty)
 {
     public static readonly ModelLaunchParams Defaults = new(
-        Temperature: 0.6,
-        TopP: 0.95,
-        TopK: 20,
-        MinP: 0.0,
-        PresencePenalty: 0.0,
-        RepetitionPenalty: 1.05);
+        Temperature: 0.6m,
+        TopP: 0.95m,
+        TopK: 20m,
+        MinP: 0.0m,
+        PresencePenalty: 0.0m,
+        RepetitionPenalty: 1.05m);
 }
